move syslog stuff into own files
